# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingDto": {
        "properties": {
          "bay_number": {
            "title": "Bay Number",
            "type": "integer"
          },
          "check_in_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check In Token"
          },
          "checked_in_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked In At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deposit_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Amount"
          },
          "discount_amount": {
            "title": "Discount Amount",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "held_at": {
            "format": "date-time",
            "title": "Held At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "merchant_id": {
            "format": "uuid",
            "title": "Merchant Id",
            "type": "string"
          },
          "merchant_service_id": {
            "format": "uuid",
            "title": "Merchant Service Id",
            "type": "string"
          },
          "payment_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method"
          },
          "payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Status"
          },
          "service_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Completed At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "merchant_id",
          "merchant_service_id",
          "bay_number",
          "status",
          "total_amount",
          "discount_amount",
          "held_at",
          "expires_at",
          "created_at"
        ],
        "title": "BookingDto",
        "type": "object"
      },
      "BookingListResponse": {
        "properties": {
          "bookings": {
            "items": {
              "$ref": "#/components/schemas/BookingDto"
            },
            "title": "Bookings",
            "type": "array"
          }
        },
        "required": [
          "bookings"
        ],
        "title": "BookingListResponse",
        "type": "object"
      },
...
      "GoldenHourResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/GoldenHourRuleDto"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules"
        ],
        "title": "GoldenHourResponse",
        "type": "object"
      },
...
      "GoldenHourUpdateRequest": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/GoldenHourRuleWrite"
            },
            "maxItems": 7,
            "title": "Rules",
            "type": "array"
          }
        },
        "title": "GoldenHourUpdateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MerchantBaysResponse": {
        "properties": {
          "bays": {
            "items": {
              "$ref": "#/components/schemas/MerchantBayDto"
            },
            "title": "Bays",
            "type": "array"
          }
        },
        "required": [
          "bays"
        ],
        "title": "MerchantBaysResponse",
        "type": "object"
      },
...
      "OpsCheckInRequest": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OpsCheckInRequest",
        "type": "object"
      },
...
      "OpsConciergeResultResponse": {
        "properties": {
          "audit_log_id": {
            "format": "uuid",
            "title": "Audit Log Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "audit_log_id"
        ],
        "title": "OpsConciergeResultResponse",
        "type": "object"
      },
      "OpsConfirmPaymentRequest": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OpsConfirmPaymentRequest",
        "type": "object"
      },
      "OpsCreateBookingRequest": {
        "properties": {
          "bay_number": {
            "minimum": 1.0,
            "title": "Bay Number",
            "type": "integer"
          },
          "merchant_id": {
            "format": "uuid",
            "title": "Merchant Id",
            "type": "string"
          },
          "merchant_service_id": {
            "format": "uuid",
            "title": "Merchant Service Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "merchant_id",
          "merchant_service_id",
          "bay_number",
          "reason"
        ],
        "title": "OpsCreateBookingRequest",
        "type": "object"
      },
...
      "SlotMaintenanceRequest": {
        "properties": {
          "bay_number": {
            "minimum": 1.0,
            "title": "Bay Number",
            "type": "integer"
          },
          "status": {
            "default": "maintenance",
            "pattern": "^(available|maintenance|closed)$",
            "title": "Status",
            "type": "string"
          },
          "time_slot": {
            "format": "date-time",
            "title": "Time Slot",
            "type": "string"
          }
        },
        "required": [
          "bay_number",
          "time_slot"
        ],
        "title": "SlotMaintenanceRequest",
        "type": "object"
      },
      "SlotMaintenanceResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "updated",
          "status"
        ],
        "title": "SlotMaintenanceResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Foundation OpenAPI contract for the TrueCare Python port.",
    "title": "TrueCare API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/bookings/{id}/arrived": {
      "post": {
        "operationId": "post_v1_bookings_by_id_arrived",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Booking Arrived",
        "tags": [
          "route-closure",
          "bookings"
        ]
      }
    },
...
    "/v1/me/bookings": {
      "get": {
        "operationId": "get_v1_me_bookings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Bookings",
        "tags": [
          "route-closure",
          "me"
        ]
      }
    },
...
    "/v1/merchants/{id}/calendar": {
      "get": {
        "operationId": "get_v1_merchants_by_id_calendar",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantBaysResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merchant Calendar",
        "tags": [
          "route-closure",
          "merchant"
        ]
      }
    },
    "/v1/merchants/{id}/calendar/maintenance": {
      "post": {
        "operationId": "post_v1_merchants_by_id_calendar_maintenance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotMaintenanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotMaintenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merchant Calendar Maintenance",
        "tags": [
          "route-closure",
          "merchant"
        ]
      }
    },
...
    "/v1/merchants/{id}/golden-hour": {
      "get": {
        "operationId": "get_v1_merchants_by_id_golden_hour",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoldenHourResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Golden Hour",
        "tags": [
          "route-closure",
          "merchant"
        ]
      },
      "put": {
        "operationId": "put_v1_merchants_by_id_golden_hour",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoldenHourUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoldenHourResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Golden Hour",
        "tags": [
          "route-closure",
          "merchant"
        ]
      }
    },
...
    "/v1/ops/bookings": {
      "post": {
        "operationId": "post_v1_ops_bookings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsCreateBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsConciergeResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Create Booking",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
    "/v1/ops/bookings/{id}/check-in": {
      "post": {
        "operationId": "post_v1_ops_bookings_by_id_check_in",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsCheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsConciergeResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Check In Booking",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
...
    "/v1/ops/data-room/{section}": {
      "get": {
        "operationId": "get_v1_ops_data_room_by_section",
        "parameters": [
          {
            "in": "path",
            "name": "section",
            "required": true,
            "schema": {
              "title": "Section",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsDataRoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Data Room",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
    "/v1/ops/evidence/upload": {
      "post": {
        "operationId": "post_v1_ops_evidence_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsEvidenceUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsConciergeResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Upload Evidence",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
    "/v1/ops/exports": {
      "post": {
        "operationId": "post_v1_ops_exports",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Export",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
    "/v1/ops/exports/{job_id}": {
      "get": {
        "operationId": "get_v1_ops_exports_by_job_id",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Export Status",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
...
    "/v1/ops/payments/{id}/confirm": {
      "post": {
        "operationId": "post_v1_ops_payments_by_id_confirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsConfirmPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsConciergeResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Confirm Payment",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
...
    "/v1/ops/reward/voucher": {
      "post": {
        "operationId": "post_v1_ops_reward_voucher",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsMintVoucherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsMintVoucherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ops Mint Reward Voucher",
        "tags": [
          "route-closure",
          "ops"
        ]
      }
    },
...
  }
}